+++ lib/rubocop/cop/style/leading_comment_space.rb
@@ -19,6 +19,15 @@ module Rubocop
             end
           end
         end
+
+        def autocorrect(comment)
+          expr = comment.loc.expression
+          b = expr.begin_pos
+          hash_mark = Parser::Source::Range.new(expr.source_buffer, b, b + 1)
+          @corrections << lambda do |corrector|
+            corrector.insert_after(hash_mark, ' ')
+          end
+        end
       end
     end
   end