+++ spec/rubocop/cop/style/space_after_colon_spec.rb
@@ -6,6 +6,8 @@ describe Rubocop::Cop::Style::SpaceAfterColon do
   subject(:cop) { described_class.new }
 
   it 'registers an offence for colon without space after it' do
+    # TODO: There is double reporting of the last colon (also from
+    # SpaceAroundOperators).
     inspect_source(cop, ['x = w ? {a:3}:4'])
     expect(cop.messages).to eq(
       ['Space missing after colon.'] * 2)
@@ -19,5 +21,10 @@ describe Rubocop::Cop::Style::SpaceAfterColon do
   it 'allows colons in strings' do
     inspect_source(cop, ["str << ':'"])
     expect(cop.messages).to be_empty
+  end
+
+  it 'auto-corrects missing space' do
+    new_source = autocorrect_source(cop, 'x = w ? {a:3}:4')
+    expect(new_source).to eq('x = w ? {a: 3}: 4')
   end
 end